Updated on 2026-06-02

+++ blockchain/src/main/java/com/tangem/blockchain/extensions/Amount.kt
@@ -4,7 +4,7 @@ import com.tangem.blockchain.common.Amount
 import java.math.BigInteger
 
 fun Amount.bigIntegerValue(): BigInteger? {
-    return this.value?.movePointRight(this.decimals.toInt())?.toBigInteger()
+    return this.value?.movePointRight(this.decimals)?.toBigInteger()
 }
 
 fun Amount.isAboveZero(): Boolean {
